+++ src/http.ts
@@ -11,7 +11,7 @@ import {
 } from "./hrana.js";
 import { SqlCache } from "./sql_cache.js";
 import { encodeBaseUrl } from "./uri.js";
-import { supportedUrlLink, extractBatchArgs } from "./util.js";
+import { supportedUrlLink } from "./util.js";
 
 export * from "./api.js";
 
@@ -72,11 +72,7 @@ export class HttpClient implements Client {
         }
     }
 
-    batch(mode: TransactionMode, stmts: Array<InStatement>): Promise<Array<ResultSet>>;
-    batch(stmts: Array<InStatement>): Promise<Array<ResultSet>>;
-    async batch(arg1: unknown, arg2: unknown = undefined): Promise<Array<ResultSet>> {
-        const {mode, stmts} = extractBatchArgs(arg1, arg2);
-
+    async batch(stmts: Array<InStatement>, mode: TransactionMode = "deferred"): Promise<Array<ResultSet>> {
         try {
             const hranaStmts = stmts.map(stmtToHrana);