+++ src/util.ts
@@ -13,22 +13,3 @@ export function transactionModeToBegin(mode: TransactionMode): string {
         throw RangeError('Unknown transaction mode, supported values are "write", "read" and "deferred"');
     }
 }
-
-export type BatchArgs = {
-    mode: TransactionMode,
-    stmts: Array<InStatement>,
-};
-
-export function extractBatchArgs(arg1: unknown, arg2: unknown): BatchArgs {
-    if (arg2 === undefined) {
-        return {
-            mode: "write",
-            stmts: arg1 as Array<InStatement>,
-        };
-    } else {
-        return {
-            mode: arg1 as TransactionMode,
-            stmts: arg2 as Array<InStatement>,
-        };
-    }
-}